fix: .int_ -> .integer
.int_ is int64, .integer is a generic jax int type; solves some warnings

+++ gymnax/environments/bsuite/deep_sea.py
@@ -179,7 +179,7 @@ class DeepSea(environment.Environment[EnvState, EnvParams]):
                     0,
                     1,
                     (self.size, self.size),
-                    dtype=jnp.int_,
+                    dtype=jnp.integer,
                 ),
                 "time": spaces.Discrete(params.max_steps_in_episode),
             }
